Generate edge id if an edge doesn't have one
