+++ kgx/transformers/nt_transformer.py
@@ -9,7 +9,7 @@ import networkx as nx
 
 from kgx import RdfTransformer
 from kgx.config import get_logger
-from kgx.utils.kgx_utils import current_time_in_millis, apply_filters
+from kgx.utils.kgx_utils import current_time_in_millis, apply_filters, generate_edge_identifiers
 
 log = get_logger()
 
@@ -63,6 +63,7 @@ class NtTransformer(RdfTransformer):
         self.dereify(self.reified_nodes)
         log.info(f"Done parsing {filename}")
         apply_filters(self.graph, self.node_filters, self.edge_filters)
+        generate_edge_identifiers(self.graph)
 
     def save(self, filename: str, output_format: str = 'nt', compression: str = None, reify_all_edges = False, **kwargs) -> None:
         """
